feat: update destop-frontend project structure
