+++ apps/desktop/src/pages/_app.tsx
@@ -1,8 +1,13 @@
+import { Providers } from "@/providers";
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return (
+    <Providers>
+      <Component {...pageProps} />
+    </Providers>
+  );
 }
 
 export default MyApp;
