+++ apps/desktop/src/views/Page.tsx
@@ -1,0 +1,19 @@
+import Head from "next/head";
+import { PropsWithChildren } from "react";
+
+const Page: React.FC<PropsWithChildren> = ({ children, ...props }) => {
+  return (
+    <div className="flex min-h-screen flex-col bg-white">
+      {/* PageMeta */}
+      <Head>
+        <title>Cap Desktop App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      {children}
+    </div>
+  );
+};
+
+export default Page;
